Backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.14.2
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.22.0
colorama==0.4.6
confection==0.1.5
cryptography==45.0.7
cymem==2.0.11
dataclasses-json==0.6.7
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
elevenlabs==2.8.0
elevenlabs-mcp==0.5.1
fastapi==0.109.2
filetype==1.2.0
Flask==2.3.2
frozenlist==1.8.0
fuzzywuzzy==0.18.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.9.0
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.0.1
langchain-google-genai==3.0.0
langchain-mcp-adapters==0.1.9
langchain-text-splitters==1.0.0
langcodes==3.5.0
langgraph==0.5.4
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.74
langsmith==0.4.8
language_data==1.3.0
Levenshtein==0.27.1
marisa-trie==1.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.12.0
mdurl==0.1.2
multidict==6.7.0
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==1.48.1
numpy==2.3.1
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.1
pdfminer.six==20250506
pillow==11.3.0
pluggy==1.6.0
preshed==3.0.10
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
PyPDF2==3.0.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
sqlparse==0.5.3
srsly==2.5.1
sse-starlette==2.4.1
starlette==0.36.3
streamlit==1.47.1
tenacity==9.1.2
thinc==8.3.6
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.27.1
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.1.1
weasel==0.4.1
websockets==15.0.1
Werkzeug==2.3.6
whitenoise==6.10.0
wrapt==1.17.3
xxhash==3.5.0
yarl==1.22.0
zstandard==0.23.0
